rework branch and link.
	modified:   include/cracker.h
	modified:   include/cracker_symbol.h
	modified:   source/cracker_arm.c
	modified:   source/cracker_main.c
	modified:   source/cracker_symbol.c
	modified:   source/cracker_thumb.c
	modified:   source/main.c

+++ include/cracker_symbol.h
@@ -27,6 +27,6 @@ enum {
 void cracker_symbol_end(symbol_p cjs, uint32_t pat, const char* name);
 void cracker_symbol_enqueue(symbol_h h2sqh, symbol_p lhs, symbol_p cjs);
 symbol_p cracker_symbol_find(cracker_p cj, symbol_h h2lhs, uint32_t pat, uint32_t mask);
-size_t cracker_symbol_intergap(cracker_p cj, symbol_p lhs, symbol_p rhs);
+int32_t cracker_symbol_intergap(cracker_p cj, symbol_p lhs, symbol_p rhs);
 void cracker_symbol_queue_log(cracker_p cj, symbol_p sqh);
 int cracker_symbol_step(cracker_p cj, symbol_p cjs);
